requirements.txt
discord.py>=1.6.0
requests>=2.22.0
